CORTX-31653: removed 3rd party deps URL from LC/update.sh
Old versions of 'cortx-build' docker image was not able to build Motr properly.
The reason was that Motr building process didn't include required installation
of 'libfabric' and 'libfabric-devel' packages as prerequisites.
In order to be able to build Motr properly it was added small workaround for
libfabric packages installation before build process is started. Latest
versions of 'cortx-build' images already contain this step as a prerequisite.
That's why described workaround is not required anymore.

Signed-off-by: Alexander Sukhachev <alexander.sukhachev@seagate.com>